requirements.txt
absl-py==0.11.0
astunparse==1.6.3
cachetools==4.1.1
certifi==2020.11.8
cffi==1.14.4
chardet==3.0.4
cmdstanpy==0.9.5
convertdate==2.3.0
cycler==0.10.0
Cython==0.29.21
ephem==3.7.7.1
fbprophet==0.7.1
gast==0.3.3
google-api-core==1.23.0
google-auth==1.23.0
google-auth-oauthlib==0.4.2
google-cloud==0.34.0
google-cloud-bigquery==2.4.0
google-cloud-bigquery-storage==2.1.0
google-cloud-core==1.4.3
google-crc32c==1.0.0
google-pasta==0.2.0
google-resumable-media==1.1.0
googleapis-common-protos==1.52.0
grpcio==1.33.2
h5py==2.10.0
holidays==0.10.3
idna==2.10
joblib==0.17.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
korean-lunar-calendar==0.2.1
libcst==0.3.14
LunarCalendar==0.0.9
Markdown==3.3.3
matplotlib==3.3.3
mypy-extensions==0.4.3
numpy==1.18.5
oauthlib==3.1.0
opt-einsum==3.3.0
pandas==1.1.4
Pillow==8.0.1
proto-plus==1.11.0
protobuf==3.14.0
pyarrow==2.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyMeeus==0.3.7
pyparsing==2.4.7
pystan==2.19.1.1
python-dateutil==2.8.1
pytz==2020.4
PyYAML==5.3.1
requests==2.25.0
requests-oauthlib==1.3.0
rsa==4.6
scikit-learn==0.23.2
scipy==1.5.4
setuptools-git==1.2
six==1.15.0
sklearn==0.0
tensorboard==2.4.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.1
tensorflow-estimator==2.3.0
termcolor==1.1.0
threadpoolctl==2.1.0
tqdm==4.53.0
typing-extensions==3.7.4.3
typing-inspect==0.6.0
urllib3==1.26.2
Werkzeug==1.0.1
wrapt==1.12.1
